# Make the hashTable argument to CompressBlock optional